separete part of logger pkg
